TS-870 Something wrong with the evacuate relevant codes in cache
Author: mohan_zl
Review: Leif, zym

git-svn-id: https://svn.apache.org/repos/asf/trafficserver/traffic/trunk@1145431 13f79535-47bb-0310-9956-ffa450edef68

+++ iocore/cache/CacheRead.cc
@@ -801,7 +801,7 @@ CacheVC::openReadStartEarliest(int event, Event * e)
     vol->begin_read(this);
 #ifdef HIT_EVACUATE
     if (vol->within_hit_evacuate_window(&earliest_dir) &&
-        (!cache_config_hit_evacuate_size_limit || doc_len <= cache_config_hit_evacuate_size_limit)) {
+        (!cache_config_hit_evacuate_size_limit || doc_len <= (uint64_t)cache_config_hit_evacuate_size_limit)) {
       DDebug("cache_hit_evac", "dir: %d, write: %d, phase: %d",
             dir_offset(&earliest_dir), offset_to_vol_offset(vol, vol->header->write_pos), vol->header->phase);
       f.hit_evacuate = 1;
@@ -1062,7 +1062,7 @@ CacheVC::openReadStartHead(int event, Event * e)
 
 #ifdef HIT_EVACUATE
     if (vol->within_hit_evacuate_window(&dir) &&
-        (!cache_config_hit_evacuate_size_limit || doc_len <= cache_config_hit_evacuate_size_limit)) {
+        (!cache_config_hit_evacuate_size_limit || doc_len <= (uint64_t)cache_config_hit_evacuate_size_limit)) {
       DDebug("cache_hit_evac", "dir: %d, write: %d, phase: %d",
             dir_offset(&dir), offset_to_vol_offset(vol, vol->header->write_pos), vol->header->phase);
       f.hit_evacuate = 1;
